+++ app/models/version.rb
@@ -3,8 +3,8 @@ class Version
 
   store_in collection: 'dataPoint', database: 'omh'
   
-  field :major, type: Integer
-  field :minor, type: Integer 
+  field :major, type: Float
+  field :minor, type: Float 
   
   embedded_in :schema_id, :inverse_of => :version
